Add Lucid and Squeeze
Both Ubuntu Lucid and Debian Squeeze are based on 2.6.32 kernels
which is basically the only thing which is supported right now.
I've tested both platforms minimally and after a few small fixes
both the SPL and ZFS build cleanly and modules can be loaded.
Add them to the platform list.